+++ Source/CesiumRuntime/Private/CesiumIonServer.cpp
@@ -43,3 +43,72 @@
 
   return Server;
 }
+
+#if WITH_EDITOR
+UCesiumIonServer*
+UCesiumIonServer::GetOrCreateForApiUrl(const FString& apiUrl) {
+  // Find an equivalent server.
+  TArray<FAssetData> CesiumIonServers;
+  FAssetRegistryModule& AssetRegistryModule =
+      FModuleManager::LoadModuleChecked<FAssetRegistryModule>("AssetRegistry");
+  AssetRegistryModule.Get().GetAssetsByClass(
+      UCesiumIonServer::StaticClass()->GetClassPathName(),
+      CesiumIonServers);
+
+  FAssetData* pFound =
+      CesiumIonServers.FindByPredicate([&apiUrl](const FAssetData& asset) {
+        UCesiumIonServer* pServer = Cast<UCesiumIonServer>(asset.GetAsset());
+        return pServer && pServer->ApiUrl == apiUrl;
+      });
+
+  if (pFound) {
+    return Cast<UCesiumIonServer>(pFound->GetAsset());
+  }
+
+  // Create a new server asset.
+  UDataAssetFactory* Factory = NewObject<UDataAssetFactory>();
+
+  UPackage* Package = nullptr;
+
+  FString PackageBasePath = TEXT("/Game/CesiumSettings/CesiumIonServers/");
+  FString PackageName;
+  FString PackagePath;
+
+  const int ArbitraryPackageIndexLimit = 10000;
+
+  for (int i = 0; i < ArbitraryPackageIndexLimit; ++i) {
+    PackageName = TEXT("FromApiUrl") + FString::FromInt(i);
+    PackagePath = PackageBasePath + PackageName;
+    Package = FindPackage(nullptr, *PackagePath);
+    if (Package == nullptr) {
+      Package = CreatePackage(*PackagePath);
+    }
+  }
+
+  if (Package == nullptr)
+    return nullptr;
+
+  Package->FullyLoad();
+
+  UCesiumIonServer* Server = Cast<UCesiumIonServer>(Factory->FactoryCreateNew(
+      UCesiumIonServer::StaticClass(),
+      Package,
+      FName(PackageName),
+      RF_Public | RF_Standalone | RF_Transactional,
+      nullptr,
+      GWarn));
+
+  Server->DisplayName = apiUrl;
+  Server->ServerUrl = apiUrl;
+  Server->ApiUrl = apiUrl;
+  Server->OAuth2ApplicationID = 190;
+
+  FAssetRegistryModule::AssetCreated(Server);
+
+  Package->FullyLoad();
+  Package->SetDirtyFlag(true);
+  UEditorLoadingAndSavingUtils::SavePackages({Package}, true);
+
+  return Server;
+}
+#endif